Clean up project for release
- Remove all log files and compilation error logs
- Remove patch scripts and fix scripts from root directory
- Remove test configuration files
- Remove obsolete archived documentation
- Remove temporary design consolidation report
- Clean up scripts directory of fix/patch/test files
- Streamline documentation structure for production release

+++ src/storage/engines/nova/mod.rs
@@ -226,7 +226,7 @@ pub fn create_vector_schema(
     if config.enable_binary {
         fields.push(Field::new(
             "vector_binary",
-            DataType::FixedSizeBinary((dimension + 7) / 8),
+            DataType::FixedSizeBinary(((dimension + 7) / 8) as i32),
             true,
         ));
     }
